refactor: AuthErrorInfo와 MemberErrorInfo 분리
