+++ src/main/java/com/ssafy/ssafsound/domain/auth/service/AuthService.java
@@ -4,12 +4,13 @@ import com.ssafy.ssafsound.domain.auth.dto.AuthenticatedMember;
 import com.ssafy.ssafsound.domain.auth.dto.CreateMemberReqDto;
 import com.ssafy.ssafsound.domain.auth.dto.CreateMemberTokensResDto;
 import com.ssafy.ssafsound.domain.auth.exception.AuthException;
-import com.ssafy.ssafsound.domain.auth.exception.MemberErrorInfo;
+import com.ssafy.ssafsound.domain.auth.exception.AuthErrorInfo;
 import com.ssafy.ssafsound.domain.auth.service.oauth.OauthProvider;
 import com.ssafy.ssafsound.domain.auth.service.oauth.OauthProviderFactory;
 import com.ssafy.ssafsound.domain.auth.service.token.JwtTokenProvider;
 import com.ssafy.ssafsound.domain.member.domain.MemberToken;
 import com.ssafy.ssafsound.domain.member.dto.PostMemberReqDto;
+import com.ssafy.ssafsound.domain.member.exception.MemberErrorInfo;
 import com.ssafy.ssafsound.domain.member.exception.MemberException;
 import com.ssafy.ssafsound.domain.member.repository.MemberTokenRepository;
 import lombok.extern.slf4j.Slf4j;
@@ -38,7 +39,7 @@ public class AuthService {
             String redirectURL = oauthProvider.getOauthUrl();
             response.sendRedirect(redirectURL);
         } catch (Exception e) {
-            throw new AuthException(MemberErrorInfo.AUTH_SERVER_ERROR);
+            throw new AuthException(AuthErrorInfo.AUTH_SERVER_ERROR);
         }
     }
 
@@ -69,8 +70,8 @@ public class AuthService {
     @Transactional(readOnly = true)
     public void validateRefreshToken(String refreshToken) {
         Long memberId = jwtTokenProvider.getMemberIdByRefreshToken(refreshToken);
-        MemberToken memberToken = memberTokenRepository.findById(memberId).orElseThrow(() -> new MemberException());
-        if(isNotEqualRefreshToken(refreshToken, memberToken.getRefreshToken())) throw new AuthException();
+        MemberToken memberToken = memberTokenRepository.findById(memberId).orElseThrow(() -> new MemberException(MemberErrorInfo.MEMBER_TOKEN_NOT_FOUND));
+        if(isNotEqualRefreshToken(refreshToken, memberToken.getRefreshToken())) throw new AuthException(AuthErrorInfo.AUTH_TOKEN_INVALID);
     }
 
     public boolean isNotEqualRefreshToken(String refreshTokenByCookie, String refreshTokenBySaved) {
