+++ src/main/java/com/ssafy/ssafsound/domain/member/exception/MemberException.java
@@ -1,6 +1,6 @@
 package com.ssafy.ssafsound.domain.member.exception;
 
-import com.ssafy.ssafsound.domain.auth.exception.MemberErrorInfo;
+import com.ssafy.ssafsound.domain.auth.exception.AuthErrorInfo;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
