+++ src/main/java/com/ssafy/ssafsound/domain/member/service/MemberService.java
@@ -2,11 +2,12 @@ package com.ssafy.ssafsound.domain.member.service;
 
 import com.ssafy.ssafsound.domain.auth.dto.AuthenticatedMember;
 import com.ssafy.ssafsound.domain.auth.exception.AuthException;
-import com.ssafy.ssafsound.domain.auth.exception.MemberErrorInfo;
+import com.ssafy.ssafsound.domain.auth.exception.AuthErrorInfo;
 import com.ssafy.ssafsound.domain.member.domain.Member;
 import com.ssafy.ssafsound.domain.member.domain.MemberRole;
 import com.ssafy.ssafsound.domain.member.domain.MemberToken;
 import com.ssafy.ssafsound.domain.member.dto.PostMemberReqDto;
+import com.ssafy.ssafsound.domain.member.exception.MemberErrorInfo;
 import com.ssafy.ssafsound.domain.member.exception.MemberException;
 import com.ssafy.ssafsound.domain.member.repository.MemberRepository;
 import com.ssafy.ssafsound.domain.member.repository.MemberRoleRepository;
@@ -35,7 +36,7 @@ public class MemberService {
 
     @Transactional
     public void saveTokenByMember(AuthenticatedMember authenticatedMember, String accessToken, String refreshToken) {
-        Member member = memberRepository.findById(authenticatedMember.getMemberId()).orElseThrow(() -> new AuthException(MemberErrorInfo.MEMBER_NOT_FOUND_BY_ID));
+        Member member = memberRepository.findById(authenticatedMember.getMemberId()).orElseThrow(() -> new MemberException(MemberErrorInfo.MEMBER_NOT_FOUND_BY_ID));
 
         MemberToken memberToken = MemberToken.builder()
                 .accessToken(accessToken)
